requirements.txt
decorator==4.0.10
ipython==5.1.0
ipython-genutils==0.1.0
m==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
pyhocon==0.3.34
pyparsing==2.1.10
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.3
traitlets==4.3.1
wcwidth==0.1.7
WebOb==1.6.2
